initial support for timed quiz
There is now a -time flag which defaults to 60 seconds. When the time is
up, the quiz is ended regardless of whether or not it is completed. The
number right/wrong is printed and the program terminates. If you finish
the quiz before the time is up, the number right/wrong is print and the
program terminates.