print the total number of questions at the end
Printing the total number of questions at the end involves reading the
whole csv file into memory at once. This is done using csv.ReadAll.

I suppose that reading the whole file into memory could be avoided. If
the user completes all of the questions, we will have a total count. If
the time expires before the user finishes, we could still finish parsing
the file in order to get the total number of questions.