initial commit for sitemap builder
It does not generate any xml yet. It only walks all of the links on a
specified site.

Factory functions have been refactored to return an error second
argument. The factory function for the sitemap builder requires some
error handle for parsing the initial url. I refactored the factory
functions of the other exercises to keep things consistent.